requirements.txt
# PyTorch生态系统
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0
# 数据处理
pandas>=1.3.0
numpy>=1.21.0

# 图像处理
Pillow>=8.3.0
opencv-python>=4.5.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0

# 机器学习
scikit-learn>=1.0.0

# 进度条
tqdm>=4.62.0

# 超参数优化 (可选)
optuna>=2.10.0

# Jupyter支持 (可选)
jupyter>=1.0.0
ipykernel>=6.0.0

# 数据科学工具 (可选)
jupyterlab>=3.0.0
